backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 支付集成
stripe==7.7.0

# 工具库
python-dotenv==1.0.0
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
